refactor(util): move util package
